Changed the function return type from size_t to int

+++ Header Files/memcmp.h
@@ -6,6 +6,6 @@
 #include <stdio.h>
 
 /*Function Declaration Section*/
-size_t my_memcmp(const char* str1, const char* str2, size_t n);
+int my_memcmp(const char* str1, const char* str2, size_t n);
 
 #endif
